requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
